requirements.txt
# Backend API
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
python-dotenv==1.0.0

# LLM & AI
groq>=0.9.0
openai==1.12.0
langchain==0.1.6
langchain-openai==0.0.5
langchain-community==0.0.19

# Vector Store & Embeddings
chromadb==0.4.22
sentence-transformers==2.3.1
huggingface-hub>=0.19.3

# Data Processing
pandas==2.2.0
numpy==1.26.3
requests==2.31.0

# NLP & Sentiment Analysis
transformers==4.37.2
torch==2.2.0

# Frontend
streamlit==1.31.0

# Utilities
requests==2.31.0
pydantic==2.6.0
pydantic-settings==2.1.0

# Testing
pytest==7.4.4
pytest-cov==4.1.0
pytest-asyncio==0.21.1
